Downloads/capstone/detector-service/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
ultralytics>=8.1.0
opencv-python-headless>=4.9.0
python-multipart>=0.0.9
numpy>=1.24.0
python-dotenv>=1.0.0
torch>=2.0.0 --index-url https://download.pytorch.org/whl/cpu
